Adjust tests to previous changes

+++ tests/sources/tools/perception/panoptic_segmentation/efficient_ps/test_efficient_ps.py
@@ -41,12 +41,13 @@ def rmdir(_dir):
 
 class TestEfficientPsLearner(unittest.TestCase):
     @classmethod
-    def setUpClass(cls) -> None:
+    def setUpClass(cls):
         print('\n\n**********************************\nTEST EfficientPS Learner\n**********************************')
 
         cls.temp_dir = os.path.join('tests', 'sources', 'tools', 'perception', 'panoptic_segmentation', 'efficient_ps',
                                     'efficient_ps_temp')
-        rmdir(cls.temp_dir)
+        if os.path.exists(cls.temp_dir):
+            rmdir(cls.temp_dir)
         os.makedirs(cls.temp_dir)
 
         # Download all required files for testing
@@ -57,7 +58,7 @@ class TestEfficientPsLearner(unittest.TestCase):
             f.extractall(cls.test_data)
 
     @classmethod
-    def tearDownClass(cls) -> None:
+    def tearDownClass(cls):
         # Clean up downloaded files
         rmdir(cls.temp_dir)
 
@@ -107,16 +108,17 @@ class TestEfficientPsLearner(unittest.TestCase):
         warnings.simplefilter('ignore', UserWarning)
 
         learner = EfficientPsLearner()
-        temp_model_path = os.path.join(self.temp_dir, 'model.pth')
+        temp_model_path = os.path.join(self.temp_dir, 'checkpoints')
         # Make sure that no model has been written to that path yet
         if os.path.exists(temp_model_path):
-            rmfile(temp_model_path)
+            rmdir(temp_model_path)
         successful = learner.save(temp_model_path)
-        self.assertTrue(os.path.exists(temp_model_path))
+        self.assertTrue(os.path.exists(os.path.join(temp_model_path, 'efficient_ps', 'efficient_ps.json')))
+        self.assertTrue(os.path.exists(os.path.join(temp_model_path, 'efficient_ps', 'model.pth')))
         self.assertTrue(successful)
-        rmfile(temp_model_path)
+        rmdir(temp_model_path)
 
-    def test_load(self):
+    def test_load_pretrained(self):
         learner = EfficientPsLearner()
         successful = learner.load(self.model_weights)
         self.assertTrue(learner._is_model_trained)
